hide details in an Internal module
