+++ recursion-schemes.cabal
@@ -80,6 +80,7 @@ library
     Data.Functor.Foldable
     Data.Functor.Base
     RecursionSchemes
+    RecursionSchemes.Internal
 
   if flag(template-haskell)
     build-depends:
